requirements.txt
Django==6.0.4
djangorestframework==3.15.1
gunicorn==22.0.0
psycopg2-binary==2.9.9
dj-database-url==2.2.0
requests==2.31.0
python-decouple==3.8
whitenoise==6.6.0
